requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
